requirements.txt
# Prime Spark AI Dependencies - Enterprise Edition

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async & Networking
aiohttp==3.9.1
httpx==0.24.1
websockets==12.0

# Memory & Caching
redis==5.0.1
pymongo==4.6.0

# Database & Storage
asyncpg==0.29.0  # PostgreSQL async driver
psycopg2-binary==2.9.9  # PostgreSQL sync driver
supabase==2.3.0
boto3==1.34.10  # For MinIO S3-compatible storage

# Streaming & Message Queue
kafka-python==2.0.2
confluent-kafka==2.3.0  # High-performance Kafka client
aiokafka==0.10.0  # Async Kafka client

# Data Pipeline
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-redis==3.4.1

# AI & ML
ollama==0.1.6
openai==1.6.1  # For API compatibility
mlflow==2.9.2  # ML experiment tracking
torch==2.1.2  # PyTorch for model deployment
transformers==4.36.2  # Hugging Face transformers
onnxruntime==1.16.3  # ONNX runtime for inference

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Task Queue
celery==5.3.4
kombu==5.3.4
flower==2.0.1  # Celery monitoring

# Analytics & Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# API & Networking
aiofiles==23.2.1
python-multipart==0.0.6
orjson==3.9.10  # Fast JSON serialization

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
ruff==0.1.8
mypy==1.7.1
pre-commit==3.6.0

# Utilities
click==8.1.7
rich==13.7.0
tenacity==8.2.3  # For retry logic
python-dateutil==2.8.2
pytz==2023.3
